Added ata_pio_write() and wrote 5 sectors (kernel) to HDD

+++ util.h
@@ -11,7 +11,7 @@ size_t strlen(const char* str);
 void memset(void *dest, uint8_t value, uint32_t byte_count);
 void outb(uint16_t port, uint8_t value);
 uint8_t inb(uint16_t port);
-void outw(uint32_t port, uint32_t value);
+void outw(uint16_t port, uint16_t value);
 uint16_t inw(uint16_t port);
 void io_wait();
